refactor(viewer): extract FEA session setup and manifest store registration from the loader
openFeaSession (fea/streaming/sessionSetup) now owns getting the mesh
GLB into the scene and the session handle around it, and
registerManifestStores (fea/streaming/manifestStores) owns pushing the
manifest's lineage, FEM concepts and groups into their stores. The
loader keeps the handle it is painting as a local resolved once, which
is what the paint modules will take as an explicit input. No behaviour
change.
